. fuseif.cpp: made CFuseSvcStat's fs_read to update its content only when the offset is zero, to avoid the inconsistent content between the contigous reads.
. propids.h, counters.cpp, counters.h, bridge.cpp, reqfwdr.cpp, rtfiles.cpp,
rpcroute.cpp, rpcroute.h, security.h: added three fast counters propMsgCount,
propMsgRespCount and propEventCount inherited from CStatCountersServer for
the 'rpcfs' file 'InConnections'.

. tools/Makefile.am: added 'x' mode to the installed python scripts.

+++ ipc/counters.cpp
@@ -99,9 +99,11 @@ gint32 IStatCounters::IncCounter(
         if( itr == m_mapCounters.end() )
         {
             if( bNegative )
-                m_mapCounters[ iProp ] = 0U;
+                m_mapCounters[ iProp ] =
+                    ( guint32 )0;
             else
-                m_mapCounters[ iProp ] = 1U;
+                m_mapCounters[ iProp ] =
+                    ( guint32 )1;
             break;
         }
